Only free the code cache repository segment when it's used
Pull request #3463 added functionality to free the code cache repository
segment when the compiler's code cache manager is destroyed. However, it
is possible to create a code cache manager that doesn't use a
repository. This commit therefore adds to check to only free the
repository segment if it's actually being used.

Signed-off-by: Leonardo Banderali <leonardo2718@protonmail.com>