Add CPU-specific tests for relative branch reachability
Add tests to each CPU class to answer whether the distance between
a source address and a target address falls within the range of
a relative, immediate branch instruction on the given CPU type.

This is useful, for example, to determine whether a trampoline is
required to reach a target method address or helper function
directly.

These functions deprecate the platform-specific macros and functions
currently in use.  They will be cleaned up once downstream projects
have been modified.

Signed-off-by: Daryl Maier <maier@ca.ibm.com>