Fix bug in arraycopy evaluator for constant byte length
With the change to generate instructions for array copy
direction test in codegen, we had a bug in case of constant
byte length array copy.
As it is constant byte length array copy, we exploit that information
in our forward direction copy path where we generate loop and encode
instruction with known constant length so in this case byte length
node won't be evaluated. In case of backward array copy we need to load
length into register where it will be used in the copying loop.
This leaves us with a diamond case where we have clobber evaluated
a byteLen node in one path but did not evaluate it in other path.
This was causing a rare bug when you have a constant length array copy
whose byte length node has more than one references and we end up using
a register set to the constant node which will have correct value in only
one path.

Signed-off-by: Rahil Shah <rahil@ca.ibm.com>