Don't delete TOC when JIT shuts down on PPC
When the compiler is shutdown and then re-initialized in the same
process, it expects to be able to reuse the same TOC as was previously
used. As such, it cannot be deleted when the JIT is shutdown. This
commit removes the code that does so from the Test Compiler, matching
the behaviour of JitBuilder.

Signed-off-by: Leonardo Banderali <leonardo2718@protonmail.com>