Acquire current CodeCache directly from CodeGenerator
Replace calls to Compilation get/setCurrentCodeCache with a direct
inquiry to the CodeGenerator.  The Compilation functions always delegate
to the CodeGenerator.  Delete the Compilation functions.

There is one instance in `CompileMethodFromDetails` where the CodeGenerator
object could be NULL (during an AOT load) and that scenario is handled
by passing a NULL CodeCache as the original Compilation function would have.

Signed-off-by: Daryl Maier <maier@ca.ibm.com>