Refine the guidance on using the TR namespace
Per the discussion at the Compiler Architecture Meeting (#3326)
and the resolution in #3306, the `TR` namespace shall be used
for the public API of the compiler component that is consumable
by downstream projects.  This PR clarifies the language and
references the umbrella issue that tracks the work to migrate
the existing public API to the `TR` namespace.

Closes #3306
Closes #2084

Signed-off-by: Daryl Maier <maier@ca.ibm.com>