Allow alternative path to merge to else block
If `elseBlock` and `mergeBlock` are identical, the user wants to create
an alternate path that after completion will jump back to the fall
through of the if test.

```
      ifBlock
       |    \
       |     \
       |   thenBlock
       |    /
       |   /
   mergeBlock/elseBlock
```

Under current implementation, `ifBlock` is constructed by splitting
`elseBlock` after prepending `ifTree` to `elseBlock`. The old
`elseBlock` will become `ifBlock`, `elseBlock` will get a new value
after split. If `mergeBlock` is the same as `elseBlock`, then it will be
the same as `ifBlock` after split. The alternate path `thenBlock` will
merge to `ifBlock` if we don't update `mergeBlock`, resulting in a
infinite loop. Fix it.

Signed-off-by: Liqun Liu <liqunl@ca.ibm.com>