Replace visit count in Inliner with node checklist
Inliner and several local transformations inside Inliner are using
visitcount to track visited nodes or call nodes that has already been
inlined and the code logic interferes with each other. Changing the
local transformations in Inliner to use NodeChecklist. Global use of
visitcount in Inliner needs more investigation and will be fixed in a
separate commit

https://github.com/eclipse/openj9/issues/3340

Signed-off-by: Yi Zhang <yizhang@ca.ibm.com>