Implement a primitive SKIP_IF macro to allow tests to be skipped
This commit introduces the SKIP_IF macro, which allows Tril-based
compiler tests to be conditionally skipped without failing or
disabling the test. This commit only intends to introduce the syntax for
skipping tests and the underlaying machinery needed for skipping test
execution. Support for features such as logging and reporting of skipped
tests is still primitive. No built-in support for platform detection is
provided.

The basic syntax is for skipping a test is:

    SKIP_IF(<condition>, <reason>);

The <condition> parameter can be any boolean expression and specifies
when a test should be skipped. The <reason> parameter must be a value
from the `SkipReason` enum; it is mainly intended for use in logging
and reporting, although the current implementation makes little use
of it.

Optionally, a "skip message" can also be streamed using the same syntax
as is used Google Test's ASSERT*() macros:

    SKIP_IF(<condition>, <reason>) << <message>;

When a test is skipped it will print a message stating the reason, the
test name, and the skip message. For example:

    [----------] 2 tests from SkipTest
    Unsupported Feature: Skipping test: SkipIfTrue
        Test is itentionally skipped to verify that skipping works
    [----------] 2 tests from SkipTest (0 ms total)

Issue #1843

Signed-off-by: Leonardo Banderali <leonardo2718@protonmail.com>