Make the Port Library a comptest dependency
The JitTest test fixture now has a static field intended to "hold" the
port library object. Two static member function are also added to the
class to initialize and shutdown the library. These are called from
the SetUp() and TearDown() methods of a global gtest environment.

One use case for having the port library available in Tril-based tests
is to skip tests conditionally. For example, using the following line
would cause a test to be skipped if it is executed on x86 and some
variable `foo` is 0:

  SKIP_IF(OMRPORT_ARCH_X86 == omrsysinfo_get_CPU_architecture() && 0 == foo);

Signed-off-by: Leonardo Banderali <leonardo2718@protonmail.com>