Unify and fix codeCacheFull functionality
There are confusingly two independent means in CodeCacheManager for
testing/setting whether the code cache is full.  Neither is behaving
as it should.

The `setCodeCacheIsFull` is never used in OMR nor any downstream project,
although the flag is tested on code memory allocation failure in
`OMR::CodeGenerator::allocateCodeMemoryInner`.

The `setCodeCacheFull` does not have an implementation in OMR (it does
in downstream projects), yet it is invoked in a number of spots in OMR.

Unify the functionality of these functions such that `setCodeCacheFull`
sets the boolean indicating that the code cache is full.  Remove the
`setCodeCacheIsFull` and `codeCacheIsFull` functions.

Signed-off-by: Daryl Maier <maier@ca.ibm.com>