Add new evaluator and refactor for ardbari opCode
1. Add ardbarEvaluator() for ardbari node.

2. Remove the hardware concurrent scavenge peephole pass
replaceGuardedLoadWithSoftwareReadBarrier(). This is OpenJ9 specific and
isn't a proper software concurrent scavenge implementation.

3. Refactor aloadHelper() by moving some of its code into separate functions so
that OMR dependant projects can make use of them.

4. Change evaluateNULLCHKWithPossibleResolve() to avoid evaluating ardbari
if it's not evaluated. ardbari has side-effects and should be evaluted
before null checked.

5. Refactor array copy evaluators. Remove referenceArrayCopyEvaluator since
most of it is J9 project specific. Remove J9 specific OOL code from
primitiveArrayCopy()

6. Rename TR_readBarrier to TR_softwareReadBarrier

7. Change handleSoftwareReadBarrier to jitSoftwareReadBarrier

8. Add a dependency parameter to generateLoadAndStoreForArrayCopy()
so that outer internal control flow's register dependency
conditions can be passed in. This makes it possible for the
load-and-store to be an inner ICF.

9. Add a register dependency conditions parameter to
generateMemToMemElementCopy() so that it can take an outer ICF's
dependency conditions and build upon it. Similar above, this makes it
possible to build mem-to-mem copy inside another ICF.

10. Add comments to the dependency parameter in
generateLoadAndStoreForArrayCopy and
generateMemToMemElementCopy.

Signed-off-by: Nigel Yu <yunigel@ca.ibm.com>