Fix GlobalVP handling of backedge constraints
Global value propagation currently handles loops using a two pass
strategy. On the first pass we do not consider constraints entering
the loop and compute a set of backedge constraints. This set of
constraints represents the information produced by an iteration of the
loop. We then process the loop again using the entry constraints and
selectively merge the backedge constraints where required. The current
def constraint merging logic considers backedge constraints only when a
def has not been seen on the current iteration of the loop. This is
insufficient - a definition must have been seen on all paths from the
loop header to the def constraint merge point to be able to ignore the
constraints on the backedge. The intuition here is that if a def has not
happened on some path from the backedge the def from a previous
iteration could reach and so we must consider it and merge them in.

This change adds a map and bitvectors of value numbers to the global VP
representation. These bitvectors track the value numbers defined on all
paths for loops and these are consulted as part of merging definitions.
Exception edges are handled conservatively in this analysis at present -
we know nothing - this can be improved if required later.

Note that failing to produce a constraint for a node is represented with
with a NULL constraint. This constraint matches the unprocessed default value.
When merging def constraints it is only safe to do this at the point of a node's
first evaluation tracking of first evaluation point is included for correctness.

Signed-off-by: Andrew Craik <ajcraik@ca.ibm.com>