Constrain the VFT when the object node is unknown
VP (value propagation) recognizes type tests in constrainIfcmpeqne,
including profiled guards, which look like this:

    ifacmpne --> slow path
      VFT
      aconst 0x... (isClassPointerConstant)

Usually, VFT here will be an aloadi <vft-symbol> from an object, and VP
will set instanceofObjectRef to the node that produces the object
reference. However, it's possible that VFT could be a direct load
instead, if e.g. the load of <vft-symbol> has been hoisted out of a
loop. In that case, instanceofObjectRef instead refers to the VFT node.
When VP attempts to apply a constraint meant for the object to the VFT
on the fallthrough path, intersection fails (because one is an object
and the other is not), and as an unintended effect of b485fd2, the
fallthrough path (i.e. the inlined body) is removed because VP believes
it is unreachable.

With this change, VP will now detect the case where instanceofObjectRef
is not an object but instead a VFT, and produce a constraint appropriate
for the VFT, preventing the guard from being spuriously folded.

Signed-off-by: Devin Papineau <devinmp@ca.ibm.com>