Fix missing omrVM in example ScavengerDelegate
This PR fixes a typo where the _omrVM member was missing from the
delegate. This error made it through because the delegate hasn't been
enabled alongside compressed references & object header.

The omrVM used to be cached in the CLI. In the ScavengerDelegate, only
keep the extensions, and just get the VM out of the extensions when
needed.

Signed-off-by: Robert Young <rwy0717@gmail.com>