Remove Ruby proof-of-concept code
An integration of OMR into the YARV Ruby VM was used as a proof-of-concept
demonstration of the versatility and relative simplicity of integrating
the technology into another language environment.

The project never matured beyond the proof-of-concept stage though, and
many of the parts that surfaced in OMR have gone stale and cannot be tested
or maintained.

Signed-off-by: Daryl Maier <maier@ca.ibm.com>