Make fatal_assertion actually "no return"
The function which actually triggers a trap,
assumeDontCallMeDirectlyImpl, cannot be marked as no-return, because
it only traps when handling fatal assertions.

Since fatal_assertion doesn't call any no-return functions, it cannot be
marked as no-return.

This PR moves the call to trap directly into fatal_assertion and
assertion, so that fatal_assertion can actually be declared "no return".

As a side effect, assumeDontCallMeDirectly is now only doing debug
printing, so it's been renamed to traceAssertionFailure, to indicate
it's more limited role in assertion failure.

Signed-off-by: Robert Young <rwy0717@gmail.com>