Make mcc_printf a variadic function macro
mcc_printf is a variable-like macro which aliases the token printf.
However, when CODECACHE_DEBUG is disabled, mcc_printf is defined to
nothing. When this happens, the arguments at each "call site" will stick
around as a parenthesized expression. These unused expressions generate
compiler warnings.

So a call like:
    mcc_printf("hello %s\n", my_name);

is transformed to the expression:
    ("hello %s\n", my_name);

This PR makes mcc_printf a variadic function-like macro that calls
printf when CODECACHE_DEBUG is on, and expanding to nothing otherwise.
By expanding the entire call to nothing, we can ensure that there are no
leftover artifacts of the disabled code.

This PR also drops the additional test against the variable
CodeCacheDebug, which isn't defined anywhere. Another define,
mcc_hashprintf, was deleted, because it simply doesn't appear in any
codebase.

As well, fix a few format-string errors in calls to mcc_printf.

Signed-off-by: Robert Young <rwy0717@gmail.com>

Fix mcc_printf formatting errors

Signed-off-by: Robert Young <rwy0717@gmail.com>