Defer bound check removal in versioner
This allows bound check removal to properly depend on privatization.

Versioner no longer rematerializes the array object expression at the
point of the bound check in buildBoundCheckComparisonsTree(). Doing so
is unnecessary, and it's incorrect in case the expression contains a
node that requiresPrivatization() while privatization cannot be
performed. Instead of rematerializing in-place beneath arrayLengthNode,
a new arrayLengthNode is created with the same subtree that would have
been effected by the in-place modifications. Uses of
boundCheckNode->getChild(boundChildIndex) are replaced with
arrayLengthNode so that they will see this substitution.

The multiple calls to performTransformation() are replaced with a single
call because there is no use in generating some but not all of the
versioning tests necessary to allow the bound check to be removed.

Versioner will temporarily refuse to remove the bound check from
BNDCHKwithSpineCHK, until spine check removal is also deferred.

Signed-off-by: Devin Papineau <devinmp@ca.ibm.com>