Defer spine check removal in versioner
This allows spine check removal to properly depend on privatization.

Versioner will now once again remove the bound check from
BNDCHKwithSpineCHK, and allow it to be considered as a spine check. The
spine check removal assumes that the spine check is a SpineCHK node. In
particular, if it was originally a BNDCHKwithSpineCHK, it must have
already been transformed into a SpineCHK (i.e. the bound check must have
been removed). For this reason, the LoopEntryPrep for the spine check
removal depends on the one for the bound check removal.

Signed-off-by: Devin Papineau <devinmp@ca.ibm.com>