Re-enable write barrier removal in versioner
Write barrier versioning is now done after the analysis that determines
whether to remove HCR guards. Write barriers can be removed if and only
if HCR guards can be removed. See the comment near the call to
buildAwrtbariComparisonsTree() in versionNaturalLoop().

Write barrier removal is also deferred, mostly for consistency with
other versioning transformations. The versioning test that allows a
write barrier to be removed cannot currently depend on privatization. It
simply tests that a loop-invariant auto containing an object reference
refers to an object that hasn't been tenured. The reference must be in
an auto because versioner only attempts this transformation when use-def
info shows that the reference is the result of an allocation visible in
the current compilation.

Signed-off-by: Devin Papineau <devinmp@ca.ibm.com>