Defer conditional folding in versioner
This allows conditional folding to properly depend on privatization.

Versioner will no longer rematerialize subexpressions of conditionalNode
in-place at the point of the conditional. Doing so is unnecessary, and
it's incorrect in general in case folding conditionalNode would require
privatization while privatization is not allowed. Instead, if part of
the subtree rooted at conditionalNode would be modified in-place,
conditionalNode is first copied (i.e. copy-on-write). The original node
is kept for use in the LoopImprovement, and for the purpose of identity
comparisons.

Signed-off-by: Devin Papineau <devinmp@ca.ibm.com>