Stop removing checkcast without a versioning test
...when performTransformation() denies permission to generate the test.
The versioning test is required in order to correctly remove the check.

The performTransformation() that could cause the test to be omitted was
in collectAllExpressionsToBeChecked(), which used to generate the test.
The checkcasts were removed in versionNaturalLoop() after
buildCheckCastComparisonsTree() had finished, where it is unknown which
(if any) of the checkcasts had versioning tests generated for them. Both
of these transformations were out of place and have been moved into
buildCheckCastComparisonsTree(), where it's clear that the checkcast is
removed if and only if the versioning test is generated.

Signed-off-by: Devin Papineau <devinmp@ca.ibm.com>