Stop removing ArrayStoreCHK without a test
...when performTransformation() denies permission to generate the
versioning test. The versioning test is required in order to correctly
remove the check.

Previously ArrayStoreCHK was removed in versionNaturalLoop() after
buildArrayStoreCheckComparisonsTree() had finished. At that position it
is unknown which (if any) of the array store checks had versioning tests
generated for them. Rather than communicate the set of checks safe to
eliminate, buildArrayStoreCheckComparisonsTree() now removes these
checks itself.

Signed-off-by: Devin Papineau <devinmp@ca.ibm.com>