Stop removing checks quite so eagerly in versioner
When generating a versioning test that happens to be the same as one
that would be generated for either a null check or a division check,
collectAllExpressionsToBeChecked() used to remove the corresponding
check (if there was one).

In preparation for deferred transformations and privatization, these
checks are no longer removed in collectAllExpressionsToBeChecked().
In order to take advantage of privatization without a guarantee that it
will succeed, it won't be possible to remove checks at this time.

All of the checks in divCheckTrees will be processed regardless, so
removing them in collectAllExpressionsToBeChecked() only served to avoid
generating duplicate versioning tests. With deferred transformations,
duplicate versioning tests won't be generated anyway.

Null checks OTOH are often skipped by versioner, so there may be null
checks left over that are guaranteed by a versioning test. These will be
dealt with when null check removal is deferred (in a later commit).

Specializer relied on the removal of these checks so that malformed
checks would not be left behind after constants were substituted. Now if
there are specialized nodes, specializer will look at each tree in the
loop and remove any such malformed checks (by transmuting to treetop).

Signed-off-by: Devin Papineau <devinmp@ca.ibm.com>