Implement versioning requiring privatization
This commit implements the framework within which versioning
transformations (including improvements to the loop and the emission of
versioning tests) will be reified and deferred. This is done because
some transformations are only correct if certain loads in the loop are
privatized. Deferral makes it possible to express the fact that the
transformation relies on privatization, and to skip the transformation
if privatization turns out not to be possible. For more details, see the
doc comment on TR_LoopVersioner.

No particular transformation is deferred in this commit. From here the
various versioning transformations can be incrementally converted.

Because the transformations are not yet deferred, and therefore do not
yet properly express their dependence on privatization, versioning
requiring privatization is still disallowed by isExprInvariant(). This
restriction will be lifted once the transformations are converted.

Signed-off-by: Devin Papineau <devinmp@ca.ibm.com>