Fold assignToHPR as false
Now that all three HPR related transformations have been deprecated
we can start cleaning up all HPR related APIs. We start with the
`assignToHPR` API which was the primary interface between global and
local RA.

This API determined whether a particular virtual register is to be
assigned to an HPR the first time the register becomes alive, or
whether when generating an instruction using such a register needs
to be upgraded to an equivalent HPR instruction.

We fold this API to always return false and all the code that
follows.

Signed-off-by: Filip Jeremic <fjeremic@ca.ibm.com>