Fold findVirtRegInHighWordRegister
This API was used to do a linear walk through the HPR register table
and find the real (HPR) register which is assigned to a particular
virtual register.

This API is in hindsight pretty expensive given the number of places
it is called in. We fold it away here and all the code that relies on
it.

Signed-off-by: Filip Jeremic <fjeremic@ca.ibm.com>