Do not generate HPR register dependencies
With HPRs almost gone out of the picture we can stop generating HPR
register dependencies since nothing can be assigned to them. Note
that the only piece we really need to keep is the KillVolHighRegs
register dependency which is used to evict volatile HPRs on 32-bit
linkage points.

This commit will greatly improve the compile time of methods as we
no longer have to traverse as many dependencies in quite a few local
RA locations.

Signed-off-by: Filip Jeremic <fjeremic@ca.ibm.com>