Restrict isAssignable to virtual registers of the same size
This is needlessly restrictive. The `registerExchange` API cannot
handle register exchanges with one 32-bit and one 64-bit register and
the code below guards against calling the `registerExchange` API in
such situations. This can definitely be relaxed and the
`registerExchange` API taught how to handle those cases. If you look
at the places this API (`isAssignable`) is used you will see we
effectively handle it there already. That code needs to be cleaned up
and consolidated into the `registerExchange` API.

I've left a TODO for this to be addressed at a later point. For now,
simply disallow the aforementioned case.

Signed-off-by: Filip Jeremic <fjeremic@ca.ibm.com>