Fold checkDependencyGroup
This API really served two purposes:

1. Calculate the available register mask
2. Various HPR related checks

We fold away the API and in the single place that it is used we
replace with with just the caluclation of the available register
map. In fact we leave a TODO here to pre-calculate this value as
it should be a compile time constant for every method compilation.
There should be no need to dynamically calculate it for every
register dependency.

The various HPR related checks set flags like `is64BitReg` and
verified that there were no duplicate HPR/GPR register dependencies.
Anecdotally this API actually consumed the most compile time in the
entire compiler (not just the codegen, even the optimizer) because of
the O(n^2) algorithm it used for each dependency list which can be
rather large given GPRS, FPRs, VRFs, and HPRs.

Signed-off-by: Filip Jeremic <fjeremic@ca.ibm.com>