Eliminate unnecessary uses of is64BitReg
During the refactoring and folding away of HPR related code we notice
quite a bit of code duplication becomes apparent particularly in the
local RA (OMRMachine.cpp). Because we do not need to special case for
HPRs a lot of the code can actually be commoned up and various
`is64BitReg` queries removed because the `if` and `else` paths are
identical.

Once the `is64BitReg` paths are folded away then it becomes apparent
that the FPR and VRF paths are now identical to that of GPR and those
can then be further folded away.

Signed-off-by: Filip Jeremic <fjeremic@ca.ibm.com>