Provide Compilation API for checking for recursive calls
Provide two APIs in the Compilation class for checking whether calling a
given method represents a recursive call to the method currently being
compiled.  One API will take a generic Symbol to analyze, and the other
will take a TR_ResolvedMethod if one is available.

Use the API throughout the code generators.

Signed-off-by: Daryl Maier <maier@ca.ibm.com>