Update the drop/restoreSMT hints for PPC
From eclipse/openj9#4797:
---
While we are at this issue, the so-called AtomicSupport drop/restore priority
primitives need to be different for AIX and pLinux, because they choose
different default hardware priority. AIX default priority is Medium (i.e. 4),
while pLinux default is Medium-Low (i.e. 3).

The current implementation is suitable to AIX (dropping to Low -- i.e. 2,
restoring to Medium). It is not suitable for pLinux (restoring to Medium which
is higher than the platform default). The suggestion for pLinux is: dropping
to Very-Low -- i.e. 1; restoring to Medium-Low which is the platform default).

The instruction to drop to Very-Low is: or r31, r31, r31;
The instruction to restore to Medium-Low is: or r6, r6, r6;
---

Signed-off-by: Dan Heidinga <daniel_heidinga@ca.ibm.com>