Fix VMRegister Adjust APIs
JitBuilder math APIs like Add, Mul, etc. require that the left and
right operands have the same type or if the left is a pointer that
the right is an Int32 or Int64. Currently a VMRegister wrapping a
int8_t * will fail when Adjust is called since _integerTypeForAdjustment
will be Int8. It is incorrect to limit the adjustBy amount to the
primitive type of the pointer since a user may want to adjust the
value by a large value. The Adjust API now takes a `size_t` amount
parameter and creates the correct constant for adjustment. This PR
includes a new test called vmregister which verifies that VMRegisters
wrapping a pInt8 work. This test failed before this PR. The
Operand[Array,Stack]Tests verify that the code continues to work
for larger types.

Signed-off-by: Charlie Gracie <charlie.gracie@gmail.com>