Fix incorrect ldiv magic values on BE platforms
When performing 64-bit signed division by a constant which is not a
power of 2, the division can be replaced with a multiplication using
some fixed point arithmetic trickery. Our previous implementation of
doing this had some hard-coded magic values for common divisors. For
unknown historical reasons, the table containing the 64-bit magic values
was actually an array of 32-bit integers which were being read as 64-bit
integers.

However, the order of 32-bit integers in the aforementioned array was
assuming that the platform was little endian. On big endian platforms,
the values of this table were being read with the two 32-bit parts of
each of the 64-bit integers swapped. As a result, 64-bit signed divides
of non-constant values by certain large constants such as 0x300000000
were being optimized using incorrect magic values on big endian
platforms.

This issue has been resolved by using an array of 64-bit integers
instead of an array of 32-bit integers read as 64-bit integers. Doing
this results in the endianness of the values being correctly handled at
compile time.

Signed-off-by: Ben Thomas <ben@benthomas.ca>