Ensure structures are updated correctly for versioned loop
In versionNaturalLoop, the loop that is being versioned and its clone
are placed into a new proper region, along with the loop's invariant
block, all preceded by a sequence of conditions.  In the process, all
edges to the invariant block are made to be edges to the sequence of
conditions at the start of the region.

However, if the original loop had an exit to the invariant block, we
end up with a situation where the new proper region structure has an
exit edge from the cloned original loop that refers to the original
invariant block, even though that block is part of that region.

Also in that same situation, the subgraph node for the cloned original
loop in the new proper region ends up pointing to itself.

Made some changes to the updating of the structures to consider this
case.

Fixes:  #3682

Signed-off-by:  Henry Zongaro <zongaro@ca.ibm.com>