Loop version support for dynamic checkcast
Currently when collecting checkcast nodes for loop invariant detection,
loop versioner asserts the cast class should always be a loadaddr. With
the upcoming CFG simplifier change, the cast class child can be dynamic.
So this restriction is removed to allow versioner to detect loop invariant
for dynamic checkcast nodes as well. The existing infrastructure seems to
be in-place to handle the new cases.

Signed-off-by: Yan Luo <Yan_Luo@ca.ibm.com>